# PERF: refactor string construction benchmark