@ngoldbaum 2023-04-04
In principle we don’t but IMO it makes the timing benchmarks more fair because it avoids a copy to numpy’s string dtype inside pandas. I have a version of this benchmark with the new string dtype I’ve been working on and avoiding that copy is *substantially* faster there.